generate token/validate token middleware ok.some specification require admin

+++ TwitterMongoDb/Controllers/UserController.cs
@@ -8,6 +8,7 @@ using System.Text;
 using TwitterMongoDb.Models;
 using TwitterMongoDb.Models.Authentication;
 using TwitterMongoDb.Services;
+using static System.Net.Mime.MediaTypeNames;
 //authentication program.cs içerisine yazıldı ancak postman da bir türlü authenticate olunamadı.ya postmanda token doğru yere yazılmıyor yada server side authentication problemi var
 namespace TwitterMongoDb.Controllers
 
@@ -30,7 +31,7 @@ namespace TwitterMongoDb.Controllers
         await _usersService.GetUsersAsync();
 
         [HttpGet("{id:length(24)}")]
-        //[Authorize]
+        [Authorize(Roles = "admin")]
         public async Task<ActionResult<User>> Get(string id)
         {
             var user = await _usersService.GetUserAsync(id);
@@ -62,12 +63,12 @@ namespace TwitterMongoDb.Controllers
                 return Unauthorized("Username geçersiz");
             }
         }
-      
+
         [HttpPost("login")]
         public async Task<IActionResult> PostLogin(UserLogin user)
         {
             var existUser = await _usersService.GetAsyncUsername(user.username);
-            if (existUser==null)
+            if (existUser == null)
             {
                 return Unauthorized("Username geçersiz");
             }
@@ -76,36 +77,12 @@ namespace TwitterMongoDb.Controllers
                 bool isPasswordValid = BCrypt.Net.BCrypt.Verify(user.password, existUser.password);
                 if (isPasswordValid)
                 {
-                    var config = new ConfigurationBuilder()
-                        .AddUserSecrets<Program>()
-                        .Build();
+                    var tokenString = JwtMiddleWares.GenerateToken(existUser);
 
-                    //jwt üretme:
-                    var secretKey = config["secretKey"]; // Özel anahtarınızı buraya ekleyin
-                    var tokenExpiration = DateTime.UtcNow.AddHours(24); // Örnek olarak 1 saatlik bir süre ekleyin
-                    var key = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(secretKey));
-                    var creds = new SigningCredentials(key, SecurityAlgorithms.HmacSha256);
-                    var claims = new[]
-                    {
-                       new Claim(ClaimTypes.Role, existUser.role),
-                    };
-                    var token = new JwtSecurityToken(config["iss"], config["aud"], claims, expires: tokenExpiration, signingCredentials: creds);
-                    var tokenString = new JwtSecurityTokenHandler().WriteToken(token);
-
-                    //Cookie olarak eklenmek istenirse:
-                    Response.Cookies.Append("jwtToken", tokenString, new CookieOptions
-                    {
-                        HttpOnly = true,
-                        Expires = tokenExpiration,
-                        Secure = true,
-                        SameSite = SameSiteMode.None,
-                        // Diğer gerekli cookie seçeneklerini ayarlayın
-                    });
-
-                    return Ok(new { token = tokenString,id=existUser.userId,name=existUser.name, username=existUser.username, message = "Giriş başarılı!" });
+                    return Ok(new { token = tokenString, id = existUser.userId, name = existUser.name, username = existUser.username, message = "Giriş başarılı!" });
                 }
             }
-            
+
             return Unauthorized("Giriş başarısız!");
         }
 
@@ -119,12 +96,15 @@ namespace TwitterMongoDb.Controllers
             {
                 return NotFound();
             }
+            if (user?.role == "admin")
+            {
+                updatedUser.userId = user.userId;
 
-            updatedUser.userId = user.userId;
+                await _usersService.UpdateUserAsync(id, updatedUser);
 
-            await _usersService.UpdateUserAsync(id, updatedUser);
-
-            return NoContent();
+                return Ok(id + " nolu user bilgileri güncellendi!");
+            }
+            return Unauthorized("Kullanıcı düzenleme yetkin yok!");
         }
 
         [HttpDelete("{id:length(24)}")]
@@ -138,9 +118,13 @@ namespace TwitterMongoDb.Controllers
                 return NotFound();
             }
 
-            await _usersService.RemoveUserAsync(id);
+            if (user?.role == "admin")
+            {
+                await _usersService.RemoveUserAsync(id);
 
-            return Ok(id +" nolu user silindi!");
+                return Ok(id + " nolu user silindi!");
+            }
+            return Unauthorized("Kullanıcı silme yetkin yok!");
         }
     }
 }
